+++ src/components/load/list.rs
@@ -1,4 +1,3 @@
-use super::*;
 use yew::prelude::*;
 use yew_router::prelude::*;
